+++ src/test/resources/fixtures/sorting/svelte/class-array-object/expected.svelte
@@ -1,0 +1,9 @@
+<button
+  class={[
+    "flex bg-red-500 p-4 text-center font-bold",
+    active && "flex p-4 font-bold",
+    { "px-2 py-0.5 text-xs tracking-wide": compact },
+  ]}
+>
+  Save
+</button>