+++ src/test/resources/fixtures/sorting/svelte/class-array-object/input.svelte
@@ -1,0 +1,9 @@
+<button
+  class={[
+    "text-center p-4 flex bg-red-500 font-bold",
+    active && "font-bold flex p-4",
+    { "tracking-wide text-xs px-2 py-0.5": compact },
+  ]}
+>
+  Save
+</button>