+++ src/test/resources/fixtures/sorting/svelte/component-class-prop/expected.svelte
@@ -1,0 +1,7 @@
+<script>
+  let props = $props()
+</script>
+
+<Button {...props} class={["flex bg-red-500 p-4 text-center font-bold", props.class]}>
+  Save
+</Button>