+++ src/test/resources/fixtures/sorting/svelte/static-template-literal/expected.svelte
@@ -1,0 +1,1 @@
+<button class={`flex bg-red-500 p-4 text-center font-bold`}>Save</button>